fix 删种

+++ app/plugins/modules/torrentremover.py
@@ -92,23 +92,12 @@ class TorrentRemover(_IPluginModule):
         for info in downloadinfos:
             if not info.DOWNLOADER or not info.DOWNLOAD_ID:
                 continue
-            # 删除标志
-            delete_flag = False
-            dl_files = self.downloader.get_files(tid=info.DOWNLOAD_ID,
-                                                 downloader_id=info.DOWNLOADER)
-            if not dl_files:
-                continue
-            for dl_file in dl_files:
-                dl_file_name = dl_file.get("name")
-                if os.path.normpath(source_file).endswith(os.path.normpath(dl_file_name)):
-                    delete_flag = True
-                    break
-            if delete_flag:
-                # 删除下载任务
-                self.__del_torrent(from_download=info.DOWNLOADER,
-                                   from_download_id=info.DOWNLOAD_ID)
+            # 删除下载任务
+            self.__del_torrent(source_file=source_file,
+                               from_download=info.DOWNLOADER,
+                               from_download_id=info.DOWNLOAD_ID)
 
-    def __del_torrent(self, from_download, from_download_id):
+    def __del_torrent(self, source_file, from_download, from_download_id):
         """
         删除下载任务
         """
@@ -131,21 +120,45 @@ class TorrentRemover(_IPluginModule):
 
             # 转种后未删除源种时，同步删除源种
             if not bool(delete_source):
-                self.info(f"删除下载任务：{from_download} - {from_download_id}")
-                self.downloader.delete_torrents(downloader_id=from_download,
-                                                ids=from_download_id)
+                # 删除标志
+                delete_flag = False
+                dl_files = self.downloader.get_files(tid=from_download_id,
+                                                     downloader_id=from_download)
+                if not dl_files:
+                    return
+                for dl_file in dl_files:
+                    dl_file_name = dl_file.get("name")
+                    if os.path.normpath(source_file).endswith(os.path.normpath(dl_file_name)):
+                        delete_flag = True
+                        break
+                if delete_flag:
+                    self.info(f"删除下载任务：{from_download} - {from_download_id}")
+                    self.downloader.delete_torrents(downloader_id=from_download,
+                                                    ids=from_download_id)
 
-        # 删除源下载任务或转种后下载任务
-        self.info(f"删除下载任务：{download} - {download_id}")
-        self.downloader.delete_torrents(downloader_id=download,
-                                        ids=download_id)
+        # 删除标志
+        delete_flag = False
+        dl_files = self.downloader.get_files(tid=download,
+                                             downloader_id=download_id)
+        if not dl_files:
+            return
+        for dl_file in dl_files:
+            dl_file_name = dl_file.get("name")
+            if os.path.normpath(source_file).endswith(os.path.normpath(dl_file_name)):
+                delete_flag = True
+                break
+        if delete_flag:
+            # 删除源下载任务或转种后下载任务
+            self.info(f"删除下载任务：{download} - {download_id}")
+            self.downloader.delete_torrents(downloader_id=download,
+                                            ids=download_id)
 
-        # 删除转种记录
-        if del_history:
-            self.delete_history(key=history_key, plugin_id=plugin_id)
+            # 删除转种记录
+            if del_history:
+                self.delete_history(key=history_key, plugin_id=plugin_id)
 
-        # 处理辅种
-        self.__del_seed(download=download, download_id=download_id)
+            # 处理辅种
+            self.__del_seed(download=download, download_id=download_id)
 
     def __del_seed(self, download, download_id):
         """
